更改字段属性

+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/PurchaseOrder.java
@@ -71,11 +71,11 @@ public class PurchaseOrder {
 
     @ApiModelProperty(value="含税采购金额")
     @JsonProperty("product_total_amount")
-    private Integer productTotalAmount;
+    private Long productTotalAmount;
 
     @ApiModelProperty(value="实物返金额")
     @JsonProperty("return_amount")
-    private Integer returnAmount;
+    private Long returnAmount;
 
     @ApiModelProperty(value="采购方式 0 配送  1.铺采直送")
     @JsonProperty("purchase_mode")
@@ -107,7 +107,7 @@ public class PurchaseOrder {
 
     @ApiModelProperty(value="赠品含税金额")
     @JsonProperty("gift_tax_sum")
-    private Integer giftTaxSum;
+    private Long giftTaxSum;
 
     @ApiModelProperty(value="公司编码")
     @JsonProperty("company_code")
